git-svn-id: https://svn.apache.org/repos/asf/incubator/jena/Import/Jena-SVN/TDB/trunk@1123376 13f79535-47bb-0310-9956-ffa450edef68

+++ src/main/java/com/hp/hpl/jena/tdb/store/bulkloader/LoaderNodeTupleTable.java
@@ -106,9 +106,9 @@ public class LoaderNodeTupleTable implements Closeable, Sync
     public void load(Node... nodes)
     {
         try {
-        count++ ;
-        monitor.dataItem() ;
-        nodeTupleTable.addRow(nodes) ;
+            count++ ;
+            monitor.dataItem() ;
+            nodeTupleTable.addRow(nodes) ;
         } catch (RuntimeException ex)
         {
             System.err.println(Iter.asString(Arrays.asList(nodes))) ;
@@ -190,7 +190,7 @@ public class LoaderNodeTupleTable implements Closeable, Sync
     {
         monitor.startIndex(label) ;
         
-        for ( int counter = 0 ; srcIter.hasNext() ; counter++ )
+        for ( long counter = 0 ; srcIter.hasNext() ; counter++ )
         {
             Tuple<NodeId> tuple = srcIter.next();
             monitor.indexItem() ;
